+++ src/app/api/models/task-definition.ts
@@ -57,10 +57,20 @@ export class TaskDefinition extends Entity {
   similarityLanguage: string = 'c';
   hasJplagReport: boolean;
   assessInPortfolioOnly: boolean;
+  requiresDiscussion: boolean;
   useResourcesForJplagBaseCode: boolean;
   lockAssessmentsToTutorialStream: boolean;
   discussionPromptsCount: number;
   overseerResourceFiles: string[] = [];
+
+  // pTargetDate: Date;
+  cTargetDate: Date;
+  dTargetDate: Date;
+  hdTargetDate: Date;
+
+  cStartDate: Date;
+  dStartDate: Date;
+  hdStartDate: Date;
 
   public readonly taskPrerequisitesCache: EntityCache<TaskPrerequisite> =
     new EntityCache<TaskPrerequisite>();