+++ src/app/api/services/task-definition.service.ts
@@ -39,6 +39,7 @@ export class TaskDefinitionService extends CachedEntityService<TaskDefinition> {
       'similarityLanguage',
       'hasJplagReport',
       'assessInPortfolioOnly',
+      'requiresDiscussion',
       {
         keys: 'targetDate',
         toEntityFn: MappingFunctions.mapDateToEndOfDay,
@@ -158,6 +159,42 @@ export class TaskDefinitionService extends CachedEntityService<TaskDefinition> {
         },
       },
       'overseerResourceFiles',
+      // {
+      //   keys: 'pTargetDate',
+      //   toEntityFn: MappingFunctions.mapDateToDay,
+      //   toJsonFn: MappingFunctions.mapDayToJson,
+      // },
+      {
+        keys: 'cTargetDate',
+        toEntityFn: MappingFunctions.mapDateToDay,
+        toJsonFn: MappingFunctions.mapDayToJson,
+      },
+      {
+        keys: 'dTargetDate',
+        toEntityFn: MappingFunctions.mapDateToDay,
+        toJsonFn: MappingFunctions.mapDayToJson,
+      },
+      {
+        keys: 'hdTargetDate',
+        toEntityFn: MappingFunctions.mapDateToDay,
+        toJsonFn: MappingFunctions.mapDayToJson,
+      },
+
+      {
+        keys: 'cStartDate',
+        toEntityFn: MappingFunctions.mapDateToDay,
+        toJsonFn: MappingFunctions.mapDayToJson,
+      },
+      {
+        keys: 'dStartDate',
+        toEntityFn: MappingFunctions.mapDateToDay,
+        toJsonFn: MappingFunctions.mapDayToJson,
+      },
+      {
+        keys: 'hdStartDate',
+        toEntityFn: MappingFunctions.mapDateToDay,
+        toJsonFn: MappingFunctions.mapDayToJson,
+      },
     );
 
     this.mapping.mapAllKeysToJsonExcept(
@@ -188,10 +225,10 @@ export class TaskDefinitionService extends CachedEntityService<TaskDefinition> {
     );
   }
 
-  public uploadOverseerResources(taskDefinition: TaskDefinition, file: File): Observable<boolean> {
+  public uploadOverseerResources(taskDefinition: TaskDefinition, file: File): Observable<string[]> {
     const formData = new FormData();
     formData.append('file', file);
-    return AppInjector.get(HttpClient).post<boolean>(
+    return AppInjector.get(HttpClient).post<string[]>(
       taskDefinition.taskOverseerResourcesUploadUrl,
       formData,
     );